requirements.txt
Django==3.1.13
django-cors-headers==3.10.0
django-environ==0.8.1
mysqlclient==2.0.3
bcrypt==3.2.0
PyJWT==2.3.0